py/requirements.txt
flask==3.0.3
pillow==10.3.0
torch==2.3.0
transformers==4.40.1
